+++ CalculatorVisitor.hpp
@@ -8,7 +8,7 @@
 class CalculatorVisitor : public BaseVisitor
 {
   public:
-    void visitNumber(int number) override;
+    void visitNumber(float number) override;
     void visitOperator(char op) override;
 
     std::unique_ptr<Ast> getRootNode();
